feat: refactor Browse components to integrate BrowseDonorVizWrapper for enhanced data visualization
